+++ src/oci/core/models/instance_configuration_instance_source_via_image_details.py
@@ -52,7 +52,7 @@ class InstanceConfigurationInstanceSourceViaImageDetails(InstanceConfigurationIn
     def boot_volume_size_in_gbs(self):
         """
         Gets the boot_volume_size_in_gbs of this InstanceConfigurationInstanceSourceViaImageDetails.
-        The size of the boot volume in GBs. Minimum value is 50 GB and maximum value is 16384 GB (16TB).
+        The size of the boot volume in GBs. The minimum value is 50 GB and the maximum value is 16384 GB (16TB).
 
 
         :return: The boot_volume_size_in_gbs of this InstanceConfigurationInstanceSourceViaImageDetails.
@@ -64,7 +64,7 @@ class InstanceConfigurationInstanceSourceViaImageDetails(InstanceConfigurationIn
     def boot_volume_size_in_gbs(self, boot_volume_size_in_gbs):
         """
         Sets the boot_volume_size_in_gbs of this InstanceConfigurationInstanceSourceViaImageDetails.
-        The size of the boot volume in GBs. Minimum value is 50 GB and maximum value is 16384 GB (16TB).
+        The size of the boot volume in GBs. The minimum value is 50 GB and the maximum value is 16384 GB (16TB).
 
 
         :param boot_volume_size_in_gbs: The boot_volume_size_in_gbs of this InstanceConfigurationInstanceSourceViaImageDetails.